Restructure notification tests into service and integration
