+++ src/test/java/de/tum/cit/aet/artemis/communication/notifications/integration/factory/TutorialGroupNotificationFactoryTest.java
@@ -1,4 +1,4 @@
-package de.tum.cit.aet.artemis.communication.notification.factory;
+package de.tum.cit.aet.artemis.communication.notifications.integration.factory;
 
 import static de.tum.cit.aet.artemis.communication.domain.notification.NotificationConstants.TUTORIAL_GROUP_DELETED_TEXT;
 import static de.tum.cit.aet.artemis.communication.domain.notification.NotificationConstants.TUTORIAL_GROUP_DELETED_TITLE;
