+++ src/test/java/de/tum/cit/aet/artemis/communication/notifications/integration/push/PushNotificationDeviceConfigurationCleanupIntegrationTest.java
@@ -1,4 +1,4 @@
-package de.tum.cit.aet.artemis.communication.service;
+package de.tum.cit.aet.artemis.communication.notifications.integration.push;
 
 import static org.springframework.test.util.AssertionErrors.assertEquals;
 
@@ -11,18 +11,16 @@ import java.util.Set;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.parallel.ResourceLock;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import de.tum.cit.aet.artemis.communication.domain.push_notification.PushNotificationDeviceConfiguration;
 import de.tum.cit.aet.artemis.communication.domain.push_notification.PushNotificationDeviceType;
 import de.tum.cit.aet.artemis.communication.repository.PushNotificationDeviceConfigurationRepository;
+import de.tum.cit.aet.artemis.communication.service.PushNotificationDeviceConfigurationCleanupService;
 import de.tum.cit.aet.artemis.core.domain.User;
 import de.tum.cit.aet.artemis.core.user.util.UserUtilService;
-import de.tum.cit.aet.artemis.shared.base.AbstractSpringIntegrationIndependentTest;
 
-@ResourceLock("pushnotifications")
-class PushNotificationDeviceConfigurationCleanupServiceTest extends AbstractSpringIntegrationIndependentTest {
+class PushNotificationDeviceConfigurationCleanupIntegrationTest extends AbstractPushNotificationIntegrationTest {
 
     @Autowired
     private PushNotificationDeviceConfigurationRepository deviceConfigurationRepository;
